Add support for parse and eval (#6)
Currently, we can only run deterministic programs via
the `driver_loop`. Programs must be entered into a prompt,
which is a little cumbersome for large programs.

Let's add support for evaluating non deterministic programs
via `parse_and_eval`.

To do this, we can define a global `try_again` function.
Within the success continuation, we can re-assign the value
of `try_again` to a function that will supply the next value.