Extend try-again to return a value (#16)
In the success continuation, instead of assigning the
next_alternative function to try_again, an anonymous
function is assigned which calls next_alternative and
returns the result.

The tests are also updated to make use of new values
being returned by calling try_again. While not necessary,
this serves to test the above functionality with the existing tests.